ready_for_printing
